+++ test-servers/configs/rfc6570-templates-http.json
@@ -3,10 +3,23 @@
     "name": "rfc6570-template-showcase",
     "version": "1.0.0"
   },
-  "resourceTemplates": [{ "preset": "rfc6570_templates" }],
-  "tools": [{ "preset": "echo" }],
+  "resourceTemplates": [
+    {
+      "preset": "rfc6570_templates"
+    }
+  ],
+  "tools": [
+    {
+      "preset": "echo"
+    }
+  ],
   "transport": {
     "type": "streamable-http",
     "port": 6603
-  }
+  },
+  "resources": [
+    {
+      "preset": "rfc6570_base"
+    }
+  ]
 }
